+++ v2/ui/src/ResourcePanel.tsx
@@ -1,38 +1,4 @@
-import * as React from 'react';
-
-
-const iconClassName = 'fa fa-cat';
-
-export const appToolbarButton = {
-    type: 'appToolbarButton',
-    factory: ({setState}: { setState: (value: any) => void }) => ({
-        iconClassName: iconClassName,
-        title: <>Example</>,
-        action: () => setState({isShown: true})
-    })
-}
-
-const ExamplePanel = ({application}: { application: any }) => {
-    return <div>
-        <h4><i className={iconClassName}/> Example Sliding Panel</h4>
-        <div className='white-box'>
-            <p>This is an example sliding panel.</p>
-            <div className='row'>
-                <div className='columns small-3'>Application</div>
-                <div className='columns small-6'>{application.metadata.name}</div>
-            </div>
-        </div>
-    </div>
-}
-
-export const appPanel = {
-    type: 'appPanel',
-    factory: ({state, setState, application}: { state: any, setState: (value: any) => void, application: any }) => ({
-        isShown: state.isShown,
-        onClose: () => setState({isShown: false}),
-        component: <ExamplePanel application={application}/>
-    })
-}
+import * as React from "react";
 
 const ExampleResourcePanel = ({
                                   state,
@@ -62,9 +28,9 @@ const ExampleResourcePanel = ({
         </div>
     </div>;
 
-export const resourcePanel = {
+export const ResourcePanel = {
     type: 'resourcePanel',
-    iconClassName: iconClassName,
+    iconClassName: 'fa fa-cat',
     title: 'Example',
     factory: ({
                   state,
@@ -77,22 +43,5 @@ export const resourcePanel = {
         component: <ExampleResourcePanel state={state} setState={setState} application={application} node={node}
                                          resource={resource}
                                          tree={tree}/>
-    })
-}
-
-const MoreButton = ({onClick}: { onClick: () => void }) => <button
-    className='argo-button argo-button--base-o argo-button--sm' onClick={onClick}>MORE</button>
-
-const ExampleStatusPanelItem = ({onClick}: { onClick: () => void }) =>
-    <div>
-        <label>EXAMPLE</label>
-        <p>This is an example.</p>
-        <MoreButton onClick={onClick}/>
-    </div>;
-
-export const statusPanelItem = {
-    type: 'appStatusPanelItem',
-    factory: ({state, setState}: { state: any, setState: (value: any) => void }) => ({
-        component: <ExampleStatusPanelItem onClick={() => setState({isShown: true})}/>
     })
 }
